refactor(compiler): Split up the const collection phase
The const collection phase is getting too large. We simplify it by breaking out i18n const collection into a separate phase.
